fix(ip-adapter): MPS-compatible fp32 stack + 512-res after smoke findings
Phase E2 smoke (shopify, 5 steps) exposed a chain of MPS + SDXL +
IP-Adapter incompatibilities. Fixes (one cohesive set, all driven by
smoke iterations):

1. scripts/train_ip_adapter.sh: prepend PYTHONPATH=${PWD} so
   accelerate launch can import the 'adapters' package (script-path
   sys.path injection only adds the script's own dir, and
   pyproject.toml's broken build-backend prevents pip install -e .).

2. scripts/train_ip_adapter.sh + configs/base.yaml: switch
   --mixed_precision and cfg.training.mixed_precision both to 'no'.
   With mixed_precision='fp16' or 'bf16' on MPS, autocast leaves
   weights in their loaded dtype (some fp16, some fp32) while casting
   activations on-the-fly, producing two failures we hit in sequence:
   (a) MPS NDArrayMatrixMultiplication assertion 'Destination NDArray
   and Accumulator NDArray cannot have different datatype' on the
   very first matmul, and (b) once that was bypassed,
   F.scaled_dot_product_attention raising 'query.dtype: float
   key.dtype: c10::Half'. accelerate's Accelerator(mixed_precision=…)
   kwarg outranks the CLI flag, so both have to agree.

3. adapters/ip_adapter/train.py: drop hardcoded
   dtype=torch.float16 from VAE and pixel_values casts (3 spots);
   add explicit torch_dtype=torch.float32 to all four
   from_pretrained calls (text_encoder_1/2, vae, unet) — SDXL ships
   its text_encoder_2 weights and madebyollin/sdxl-vae-fp16-fix
   primarily as fp16, so without an explicit dtype the loader yields
   mixed-precision frozen models that re-trigger the K/V mismatch.

4. adapters/ip_adapter/model.py: load CLIPVisionModelWithProjection
   with torch_dtype=torch.float32; cast ip_key/ip_value to
   query.dtype after to_k_ip/to_v_ip (defensive — mirrors the
   existing main-attention 'hidden_states.to(query.dtype)' pattern
   at model.py:135 and prevents a future regression once mixed
   precision is workable on MPS again).

5. configs/ip_adapter/{shopify,etsy,ebay}.yaml: image_size
   768 -> 512. The fp32 path measured ~28 s per effective step at
   768, projecting ~24 h per platform — over budget for the 3-day
   total. CLIP image encoder input is 336x336 regardless, so the
   per-platform aesthetic signal is unaffected; reducing the
   diffusion path resolution costs minimal quality and brings
   each platform to ~6 h.

Smoke (post-fix) result: 5 effective steps + 1 final validate
completed cleanly; train.log shows two identical
'step=5 val_loss=0.070383 wall=139s n=44' entries (in-loop +
post-loop with deterministic-seed RNG); checkpoint-5/ and final/
both contain image_proj_model.pt + ip_attn_processors.pt.

+++ adapters/ip_adapter/train.py
@@ -183,7 +183,7 @@ def _validate(
         wall_start = time.time()
 
         for batch in val_loader:
-            pixel_values = batch["pixel_values"].to(accelerator.device, dtype=torch.float16)
+            pixel_values = batch["pixel_values"].to(accelerator.device)
             clip_pixel_values = batch["clip_pixel_values"].to(accelerator.device, dtype=torch.float32)
 
             latents = vae.encode(pixel_values).latent_dist.sample(generator=gen)
@@ -264,12 +264,15 @@ def train(cfg_path: str) -> None:
     )
 
     # ---- Models ----
+    # Force fp32 explicitly — some SDXL weight files default to fp16 storage
+    # (e.g. text_encoder_2 as `*.fp16.safetensors`, madebyollin VAE is fp16-first),
+    # which mixes dtypes in cross-attention on MPS and raises at SDPA.
     tokenizer_1 = CLIPTokenizer.from_pretrained(cfg.model.base, subfolder="tokenizer")
     tokenizer_2 = CLIPTokenizer.from_pretrained(cfg.model.base, subfolder="tokenizer_2")
-    text_encoder_1 = CLIPTextModel.from_pretrained(cfg.model.base, subfolder="text_encoder")
-    text_encoder_2 = CLIPTextModelWithProjection.from_pretrained(cfg.model.base, subfolder="text_encoder_2")
-    vae = AutoencoderKL.from_pretrained(cfg.model.vae)
-    unet = UNet2DConditionModel.from_pretrained(cfg.model.base, subfolder="unet")
+    text_encoder_1 = CLIPTextModel.from_pretrained(cfg.model.base, subfolder="text_encoder", torch_dtype=torch.float32)
+    text_encoder_2 = CLIPTextModelWithProjection.from_pretrained(cfg.model.base, subfolder="text_encoder_2", torch_dtype=torch.float32)
+    vae = AutoencoderKL.from_pretrained(cfg.model.vae, torch_dtype=torch.float32)
+    unet = UNet2DConditionModel.from_pretrained(cfg.model.base, subfolder="unet", torch_dtype=torch.float32)
     noise_scheduler = DDPMScheduler.from_pretrained(cfg.model.base, subfolder="scheduler")
 
     adapter = IPAdapterSDXL(
@@ -330,7 +333,7 @@ def train(cfg_path: str) -> None:
         adapter, optimizer, dataloader, lr_scheduler, val_loader
     )
     train_log = output_dir / "train.log"
-    vae = vae.to(accelerator.device, dtype=torch.float16)
+    vae = vae.to(accelerator.device)
     text_encoder_1 = text_encoder_1.to(accelerator.device)
     text_encoder_2 = text_encoder_2.to(accelerator.device)
 
@@ -341,7 +344,7 @@ def train(cfg_path: str) -> None:
         adapter.train()
         for batch in dataloader:
             with accelerator.accumulate(adapter):
-                pixel_values = batch["pixel_values"].to(accelerator.device, dtype=torch.float16)
+                pixel_values = batch["pixel_values"].to(accelerator.device)
                 # CLIP encoder expects fp32; keep on same device
                 clip_pixel_values = batch["clip_pixel_values"].to(accelerator.device, dtype=torch.float32)
